refactor(exception): 디렉토리 구조 변경
- Global 폴더로 전역으로 사용하는 common, config, exception 폴더를 이동 했습니다.
