--- src/main/java/com/delivery/justonebite/exception/test.txt
@@ -1,1 +1,0 @@
-파일 업로드용 txt
