+++ src/main/java/com/delivery/justonebite/order/service/OrderService.java
@@ -1,0 +1,4 @@
+package com.delivery.justonebite.order.service;
+
+public class OrderService {
+}
